CDP #117 - Moving "getNestedValue" function from "core-data" to "shared" package
